// src/enums/e-possible-sub-stats.ts
export enum EPossibleSubStats {
  atk = "atk",
  atkPercentage = "atkPercentage",
  hp = "hp",
  hpPercentage = "hpPercentage",
  def = "def",
  defPercentage = "defPercentage",
  spd = "spd",
  breakEffect = "breakEffect",
  critRate = "critRate",
  critDmg = "critDmg",
  effectHitRate = "effectHitRate",
}